requirements.txt
alembic==1.9.3
anyio==3.6.2
click==8.1.3
ecdsa==0.18.0
fastapi==0.90.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.4
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.23.0
typing_extensions==4.4.0
uvicorn==0.20.0
